Complete compile time dependency change. Add travis runtime test subroutine
