+++ dsn_test.go
@@ -1,7 +1,6 @@
 package gosnowflake
 
 import (
-	"fmt"
 	"net/url"
 	"reflect"
 	"testing"
@@ -15,9 +14,6 @@ type tcParseDSN struct {
 }
 
 func TestParseDSN(t *testing.T) {
-
-	privKeyPKCS8 := generatePKCS8StringSupress(TestPrivKey)
-	privKeyPKCS1 := generatePKCS1String(TestPrivKey)
 
 	testcases := []tcParseDSN{
 		{
@@ -122,24 +118,7 @@ func TestParseDSN(t *testing.T) {
 			},
 			err: nil,
 		},
-		{
-			dsn: fmt.Sprintf("u:p@a.snowflake.local:9876?account=a&protocol=http&authenticator=SNOWFLAKE_JWT&privateKey=%v", privKeyPKCS8),
-			config: &Config{
-				Account: "a", User: "u", Password: "p",
-				Authenticator: authenticatorJWT, PrivateKey: TestPrivKey,
-				Protocol: "http", Host: "snowflake.local", Port: 9876,
-			},
-			err: nil,
-		},
-		{
-			dsn: fmt.Sprintf("u:p@a.snowflake.local:9876?account=a&protocol=http&authenticator=SNOWFLAKE_JWT&privateKey=%v", privKeyPKCS1),
-			config: &Config{
-				Account: "a", User: "u", Password: "p",
-				Authenticator: authenticatorJWT, PrivateKey: TestPrivKey,
-				Protocol: "http", Host: "snowflake.local", Port: 9876,
-			},
-			err: &SnowflakeError{Number: ErrCodePrivateKeyParseError},
-		},
+
 		{
 			dsn: "u:p@a?database=d&jwtTimeout=20",
 			config: &Config{
@@ -251,6 +230,9 @@ func TestParseDSN(t *testing.T) {
 			err: nil,
 		},
 	}
+
+	testcases = addParseDSNTest(testcases)
+
 	for i, test := range testcases {
 		glog.V(2).Infof("#%v\n", i)
 		cfg, err := ParseDSN(test.dsn)
