+++ optional_go10-_test.go
@@ -1,0 +1,11 @@
+// +build !go1.10
+
+// This files contains variable or function of test cases that we want to run for go version <= 1.10
+
+// See header comments on optional_go10+_test.go
+
+package gosnowflake
+
+func addParseDSNTest(parseDSNTests []tcParseDSN) []tcParseDSN {
+	return nil
+}
